requirements.txt
pandas
numpy
requests
scikit-learn
python-dotenv
schedule
streamlit
